update teams dashboard
