+++ resources/js/langs/en/nav.json
@@ -9,5 +9,6 @@
     "settings": "Settings",
     "logout": "Logout",
     "login": "Login",
-    "signup": "Signup"
+    "signup": "Signup",
+    "teams": "Teams"
 }
