+++ resources/js/views/Teams/MyTeams.vue
@@ -1,0 +1,17 @@
+<template>
+    <section>
+        <div>
+            <p>My Teams</p>
+        </div>
+    </section>
+</template>
+
+<script>
+export default {
+    name: 'MyTeams'
+}
+</script>
+
+<style scoped>
+
+</style>
