--- resources/js/views/settings/Teams/Default.vue
@@ -1,35 +1,0 @@
-<template>
-    <div>
-        <p class="mb1">I want to...</p>
-
-        <div class="flex">
-            <button
-                class="button is-medium is-info mr1"
-                @click="goto('CreateTeam')"
-                v-show="remaining > 0"
-            >Create a new team</button>
-
-            <button class="button is-medium is-primary" @click="goto('JoinTeam')">Join an existing team</button>
-        </div>
-    </div>
-</template>
-
-<script>
-export default {
-    name: 'Default',
-    props: ['remaining'],
-    methods: {
-        /**
-         * Change component type
-         */
-        goto (type)
-        {
-            this.$store.commit('teamComponent', type);
-        }
-    }
-}
-</script>
-
-<style scoped>
-
-</style>
